feat(#7): set up route and route stack
