+++ package.json
@@ -10,6 +10,8 @@
     "eslint": "eslint src/* --no-error-on-unmatched-pattern"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.6",
+    "@react-navigation/native-stack": "^6.2.5",
     "expo": "~43.0.0",
     "expo-status-bar": "~1.1.0",
     "firebase": "^9.5.0",
