+++ src/App.tsx
@@ -1,34 +1,17 @@
 import React from 'react';
 import { registerRootComponent } from 'expo';
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
-import BottomTab, { BottomTabView } from './component/BottomTabs';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import Home from './page/Home';
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+const Stack = createNativeStackNavigator();
 
 const Index: React.FC = () => (
-  <View style={styles.container}>
-    <Text>Open up App.tsx to start working on your app!</Text>
-    <BottomTab>
-      <BottomTabView title="test1">
-        <Text>heyhey123</Text>
-      </BottomTabView>
-      <BottomTabView title="test2">
-        <Text>heyhey234</Text>
-      </BottomTabView>
-      <BottomTabView title="test3">
-        <Text>heyhey2345</Text>
-      </BottomTabView>
-    </BottomTab>
-    <StatusBar />
-  </View>
+  <NavigationContainer>
+    <Stack.Navigator initialRouteName="Home">
+      <Stack.Screen name="Home" component={Home} />
+    </Stack.Navigator>
+  </NavigationContainer>
 );
 
 registerRootComponent(Index);
